Fix MBF static buffer bug
The array m_freeBufferBitmap was larger than necessary when
ERPC_DEFAULT_BUFFERS_COUNT is divisible by 8. This is not a bug, but it
is a bit inefficient. The memset of m_freeBufferBitmap was shorter than
the length of the array which had the effect of pre-allocating
(ERPC_DEFAULT_BUFFERS_COUNT % 8) of the buffers.

Thanks to Peter Hinson for identifying this issue.